# sevfruit-supplier-service/api/api-spec.yaml
...
openapi: 3.1.0
info:
  title: Сервис учёта поставок
  description: |
    API сервиса учёта поставок.
    Используется для внесения информации о поставщиках, продукции, ценах на продукцию на каждый период поставок и поставках продукции.
    Предусмотрено создание отчёта о поставках продукции поставщиками в заданном периоде.
    
    ## Авторизация  
    Для определения прав доступа к методам API используется авторизация API key. Ключ доступа передаётся в заголовке *Api-Key*.  
    
    ## Ошибки  
    API использует стандартные HTTP коды для сообщения об успешности или не успешности вызова API. В случае неуспеха, теле ответа
    содержится JSON с информацией об ошибке.
  version: '1.0'
  license:
    name: Unlicense
    url: https://opensource.org/license/unlicense/
servers:
  - url: http://localhost:8080
    description: Локальный сервер
security:
  - apikey: []
...
paths:
...
  /product:
    get:
      operationId: productGetAll
      summary: Список всех видов продукции
      description: Возвращает список всех видов продукции
      tags:
      - Product
      responses:
        "200":
          description: OK
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
              example:
              - id: 1
                name: Яблоки красные
              - id: 2
                name: Яблоки жёлтые
              - id: 3
                name: Яблоки зелёные
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /product
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /product
    post:
      operationId: productAdd
      summary: Добавить новый вид продукции
      description: Добавляет новый вид продукции
      tags:
      - Product
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
            example:
              name: Яблоки белые
      responses:
        "201":
          description: OK
          headers:
            Location:
              description: URL нового вида продукции
              schema:
                type: string
                format: uri
              example: http://localhost:8080/product/7
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
              example:
                id: 7
                name: Яблоки белые
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /product
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /product
  /product/{product_id}:
    get:
      operationId: productGet
      summary: Заданный вид продукции
      description: Возвращает заданный вид продукции
      tags:
      - Product
      parameters:
        - name: product_id
          in: path
          description: ID вида продукции
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
              example:
                id: 1
                name: "Яблоки красные"
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /product/1
        "404":
          description: Вид продукции не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 404
                error: Not Found
                message: Вид продукции 11 не найден
                path: /product/11
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /product/1
  /period:
    get:
      operationId: periodGetAll
      summary: Список всех периодов поставок
      description: Возвращает список всех периодов поставок
      tags:
      - Period
      responses:
        "200":
          description: OK
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Period"
              example:
              - id: 1
                name: "2020"
              - id: 2
                name: "2021"
              - id: 3
                name: "2022"
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /period
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /period
    post:
      operationId: periodAdd
      summary: Добавить новый период поставок
      description: Добавляет новый период поставок
      tags:
      - Period
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Period"
            example:
              name: "2023"
      responses:
        "201":
          description: OK
          headers:
            Location:
              description: URL нового периода поставок
              schema:
                type: string
                format: uri
              example: http://localhost:8080/period/4
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Period"
              example:
                id: 4
                name: "2023"
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /period
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /period
  /period/{period_id}:
    get:
      operationId: periodGet
      summary: Заданный период поставок
      description: Возвращает заданный период поставок
      tags:
      - Period
      parameters:
        - name: period_id
          in: path
          description: ID периода
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Period"
              example:
                id: 3
                name: "2022"
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /period/3
        "404":
          description: Период не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 404
                error: Not Found
                message: Период 33 не найден
                path: /period/33
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /period/3
  /period/{period_id}/report:
    get:
      operationId: periodGetSupplierProductReport
      summary: Отчёт о поступлении видов продукции
      description: Возвращает отчёт о поступлении видов продукции по поставщикам с итогами по весу и стоимости
      parameters:
        - name: period_id
          in: path
          description: ID периода поставок
          required: true
          schema:
            type: integer
      tags:
      - Period
      responses:
        "200":
          description: OK
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierProductsReport"
              example:
              - supplier:
                  id: 3
                  name: S3
                products:
                - product:
                    id: 1
                    name: Яблоки красные
                  totalQuantity: 2.0
                  totalValue: 220000.0
                - product:
                    id: 4
                    name: Груши жёлтые
                  totalQuantity: 3.0
                  totalValue: 450000.0
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /period/3/report
        "404":
          description: Период не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 404
                error: Not Found
                message: Период 33 не найден
                path: /period/33/report
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /period/3/report                
...
components:
  schemas:
    Supplier:
      type: object
      title: Поставщик
      description: Описание поставщика
      properties:
        id:
          type: integer
          format: int32
          description: ID поставщика
          readOnly: true
        name:
          type: string
          description: Наименование поставщика, уникальное для данного сервиса
      required:
      - name
    Product:
      type: object
      title: Вид продукции
      description: Описание вида продукции
      properties:
        id:
          type: integer
          format: int32
          description: ID вида продукции
          readOnly: true
        name:
          type: string
          description: Наименование вида продукции, уникальное для данного сервиса
      required:
      - name
    Period:
      type: object
      title: Период поставок
      description: Описание периода поставок
      properties:
        id:
          type: integer
          format: int32
          description: ID периода поставок
          readOnly: true
        name:
          type: string
          description: Наименование периода поставок, уникальное для данного сервиса
      required:
      - name
...
    SupplierProductsReport:
      type: object
      title: Отчёт о поставках видов продукции
      description: Отчёт о поставках видов продукции с итогами по количеству и стоимости
      properties:
        supplier:
          $ref: "#/components/schemas/Supplier"
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductTotal"
      required:
      - supplier
    ProductTotal:
      type: object
      title: Вид продукции с итогами по количеству и стоимости
      description: Вид продукции из отчёта с итогами по количеству и стоимости
      properties:
        product:
          $ref: "#/components/schemas/Product"
        totalQuantity:
          type: number
          format: float
          description: Общее количество в тоннах
        totalValue:
          type: number
          format: float
          description: Общая стоимость
      required:
      - product
      - totalQuantity
      - totalValue
    ErrorResponse:
      type: object
      title: Ошибка
      description: Описание ошибки
      properties:
        timestamp:
          type: string
          description: Дата и время
          readOnly: true
        status:
          type: integer
          description: HTTP код ошибки
          readOnly: true
        error:
          type: string
          description: Текстовое описание HTTP кода ошибки
          readOnly: true
        message:
          type: string
          description: Сообщение об ошибке
          readOnly: true
        path: 
          type: string
          description: Путь запроса
          readOnly: true
      required:
      - timestamp
      - status
      - error
  securitySchemes:
    apikey:
      type: apiKey
      in: header
      name: Api-Key
      description: Авторизация с помощью API key